Refactor getSysUserId into a Separate class
