+++ src/main/java/org/openelisglobal/common/rest/provider/PatientSearchRestController.java
@@ -6,12 +6,12 @@ import java.util.ArrayList;
 import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import org.apache.commons.validator.GenericValidator;
+import org.openelisglobal.SysUserId.SysUserId;
 import org.openelisglobal.common.provider.query.PatientSearchResults;
 import org.openelisglobal.common.provider.query.PatientSearchResultsForm;
 import org.openelisglobal.common.provider.query.workerObjects.PatientSearchLocalAndExternalWorker;
 import org.openelisglobal.common.provider.query.workerObjects.PatientSearchLocalWorker;
 import org.openelisglobal.common.provider.query.workerObjects.PatientSearchWorker;
-import org.openelisglobal.common.rest.BaseRestController;
 import org.openelisglobal.common.rest.util.PatientSearchResultsPaging;
 import org.openelisglobal.common.util.ConfigurationProperties;
 import org.openelisglobal.common.util.ConfigurationProperties.Property;
@@ -35,7 +35,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
 @RequestMapping(value = "/rest/")
-public class PatientSearchRestController extends BaseRestController {
+public class PatientSearchRestController extends SysUserId {
 
     @Autowired
     SampleService sampleService;
