+++ src/main/java/org/openelisglobal/notifications/rest/NotificationRestController.java
@@ -10,6 +10,7 @@ import javax.servlet.http.HttpServletRequest;
 import nl.martijndwars.webpush.PushService;
 import org.apache.http.HttpResponse;
 import org.bouncycastle.jce.provider.BouncyCastleProvider;
+import org.openelisglobal.SysUserId.SysUserId;
 import org.openelisglobal.login.valueholder.UserSessionData;
 import org.openelisglobal.notifications.dao.NotificationDAO;
 import org.openelisglobal.notifications.dao.NotificationSubscriptionDAO;
@@ -31,7 +32,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RequestMapping("/rest")
 @RestController
-public class NotificationRestController {
+public class NotificationRestController extends SysUserId {
 
     private final NotificationDAO notificationDAO;
     private final SystemUserService systemUserService;
@@ -211,14 +212,4 @@ public class NotificationRestController {
         return ResponseEntity.ok().body("Unsubscribed successfully");
     }
 
-    protected String getSysUserId(HttpServletRequest request) {
-        UserSessionData usd = (UserSessionData) request.getSession().getAttribute(USER_SESSION_DATA);
-        if (usd == null) {
-            usd = (UserSessionData) request.getAttribute(USER_SESSION_DATA);
-            if (usd == null) {
-                return null;
-            }
-        }
-        return String.valueOf(usd.getSystemUserId());
-    }
 }
