+++ src/main/java/org/openelisglobal/program/controller/pathology/PathologyController.java
@@ -7,7 +7,8 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import javax.servlet.http.HttpServletRequest;
-import org.openelisglobal.common.rest.BaseRestController;
+
+import org.openelisglobal.SysUserId.SysUserId;
 import org.openelisglobal.program.bean.PathologyDashBoardCount;
 import org.openelisglobal.program.service.PathologyDisplayService;
 import org.openelisglobal.program.service.PathologySampleService;
@@ -27,7 +28,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class PathologyController extends BaseRestController {
+public class PathologyController extends SysUserId {
 
     @Autowired
     private PathologySampleService pathologySampleService;
