+++ src/main/java/org/openelisglobal/qaevent/controller/rest/ViewNonConformEventsRestController.java
@@ -5,6 +5,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
+
+import org.openelisglobal.SysUserId.SysUserId;
 import org.openelisglobal.common.services.DisplayListService;
 import org.openelisglobal.common.util.DateUtil;
 import org.openelisglobal.login.valueholder.UserSessionData;
@@ -29,7 +31,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class ViewNonConformEventsRestController {
+public class ViewNonConformEventsRestController extends SysUserId{
 
     @Autowired
     private NCEventService ncEventService;
@@ -121,14 +123,5 @@ public class ViewNonConformEventsRestController {
         }
     }
 
-    protected String getSysUserId(HttpServletRequest request) {
-        UserSessionData usd = (UserSessionData) request.getSession().getAttribute(USER_SESSION_DATA);
-        if (usd == null) {
-            usd = (UserSessionData) request.getAttribute(USER_SESSION_DATA);
-            if (usd == null) {
-                return null;
-            }
-        }
-        return String.valueOf(usd.getSystemUserId());
-    }
+    
 }
